x2goserver/bin/x2gostartagent: checking for the exit status of a variable assignment isn't helping us in finding out if x2goagent is still running.
Even otherwise, the exit status would be useless since we spawn
x2goagent in the background.

Better check if the PID still exists and error out if it's gone.

There certainly is potential for race conditions (i.e., x2goagent
started, then exited, then another process took its place). On Linux,
this situation is unlikely to happen, since PIDs are only recycled once
the maximum PID value is reached, meaning that enough processes to
overflow the PID + to get to the original PID value need to be spawned
in a very short amount of time.

Unlikely, but not impossible.

Cherry-picked from release/4.0.1.x branch.